requirements.txt
Django==4.2
djangorestframework==3.14.0
requests==2.28.1
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
tensorflow==2.12.0
keras==2.12.0
gunicorn==20.1.0
python-dotenv==1.0.0
